adts-inference/requirements.txt
keras==2.9.0
numpy==1.21.5
pandas==1.4.3
pyod==1.0.3
scikit-learn==1.1.1
sktime==0.13.0
tensorflow-cpu==2.9.1